+++ src/mailmanclient/docs/NEWS.rst
@@ -8,6 +8,7 @@ NEWS for mailmanclient
 * Add a ``mail_host`` parameter to ``get_list_page`` and ``find_lists`` to
   support filtering the response by a list domain.
 * URL encode values in URL which are url unsafe. (Closes #44)
+* Add support to mass unsubscribe memebrs from a Mailing List. (Closes #43)
 
 
 3.2.2 (2019-02-09)